langchain-examples/pyproject.toml
[project]
name = "langchain-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.21",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langchain[google-genai]>=1.2.13",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
